PWS-105 Prevend data loss when workflow is removed.
When a workflow was removed the model was not correctly cleaned up.
As the result when periodic save happens it overrides the model of newly
selected workflow with the model of removed workflow.

Now when a workflow is removed we import the model of newly selected workflow
into the studio.