Simplified property view in the web studio.
Now we support two property views for jobs and tasks:
- the simple one with the minimum set of fields that are often used (activated by default)
- the details view with all posibble jobs/tasks properties

The studio allows to switch between these views and remembers which one is choosen for each workflow.

For the simple view new fields are introduced releated to host criteria for tasks:
- host name
- operating system
- free memory
- exclusivity

These fields are replaced by selection scripts automatically. Values of
these extra fields are stored in the generic information of tasks. It makes possible to
export/import jobs saving values of these custom fields.

When switching from detailed to simple view selection scripts are regenerated and
all custom selection scripts are lost. There is a warning dialog about it.

Change-Id: I55eb25f78d963b2ee29849197a0c4c00ed8a3b42