Web studio simplified
What was added:
- simple forms for jobs and tasks
- a switch between simple and adwanced view
- some criteria to tasks that transform into selection scripts or topology

Simple form is not permanent yet (will be reset after reload) as there is not transformation from the model to this simple form.
E.g. selection scripts that check the host name should be transformed into just one text field.

It looks ugly as design has not been improved yet.