PWS-156 Fix the exception from jquery draggable library on task "drag&drop".
The problems was related to reloading of palette view every time a task is
dropped (in template editing mode). So at the moment of the task drop
the menu is recreated in the pallete view and all droppable elements
are reinitialized.

As a solution palette view is rendered now when workflow is saved.