Bump up ui-tests libs versions and update ui tests
UI-tests were broken due to compatibility issues between Firefox and ui-tests libraries (nightwatch.js and Selenium). Also, ui-tests were not reflecting all recent changes to the studio (default mode is detailed view) so they have been updated in that regard. Selenium jar library is versioned because this jar is not available on maven.