Remove the test if /usr is entered by the script.
This check is removed for the sake of working studio test. So that we can finally rely on them. Nevertheless, the test cases need to be improved but selenium/nightwatch doesn't show deterministic behavior. This check executes successfully on local machines and the trydev interface satifsies this requirement. But somehow this check doesn't work on the server side, maybe because of a fake screen or whatever. We need to investigate that.