Append template content to document object at first during rendering
Template content associated to login view was not always appended to the
document object while rendering the login view with Backbone.js. Also, binding
was too late for some of the cases that were already handled. As a consequence,
the event that is defined for the Submit button was not registered by
Backbone.js and thus not triggered.

This patch fixes ow2-proactive/studio#312.