Fix failing test
The testcase Workflow with job variable was failing because it was actually login out before doing the rest of the checks. In a local environment this doesnt happen because a notification appears on top of the log out button but with latency, the notification has enough time to disappear so the way is clear to log out. This fix is to remove the unecessary click to log out as the test is not finished.