Update the Scheduler mockup
The latest improvements to the Catalog broke compatibility with the tests and the Scheduler mockup. I fixed it so we can continue to use this mockup when working on the studio. I also added a new grunt task so it's easier to run, just use the following command: grunt mock:run. It can also be used to test new development with: grunt clean build mock:run which will expose the studio without having to run a scheduler in the backend. Everything should work as if except the submission of workflows (obviously)