Use textarea in variable values
When entering long values in variables, as the input element width is not resizable, it is complicated to view the value entered or spot mistakes.

This changes the input field used for variable-like values (Job and Task variables, Generic information, and job submission) from text to textarea, allowing to resize the field and thus better view the entered value.

Additionally:
 - make all textarea resizable vertically only (resizing horizontally creates issues on the studio).
 - change the label "Number of Automatic Restarts", which is incorrect, to "Number of Execution Attempts"